refactor: alter column

+++ database/migrations/2025_08_17_185945_add_end_conversation_time_to_transcripts_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('transcripts', function (Blueprint $table) {
-            $table->bigInteger('end_conversation_time')->nullable()->after('conversation');
+            $table->integer('end_conversation_time')->nullable()->after('conversation');
         });
     }
